requirements.txt
fastapi
uvicorn
fpdf2
aiofiles
